+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [2.5.0]
+- Fixed cursor position (tks @benz93chung)
+- Fixed double value not reseting when clearing content from text field
+- added init value
+- added precision to double value
+
 ## [2.0.0]
 - Migrate to null-safety
 ## [1.0.1] - [1.0.2] - [1.0.3] - 2019-10-07.